Permite valores decimais com vírgula no campo abatimento_mm
- models.py: IntegerField → DecimalField(max_digits=6, decimal_places=2)
- forms/produtos.py: adiciona DecimalVirgula que aceita vírgula como
  separador decimal (pt-BR) nos formulários de Perfil, PerfilPuxador e Puxador
- migration 0036: aplica a alteração no banco

+++ portas/models.py
@@ -40,7 +40,7 @@ class ProdutoBase(AtivoModel):
     descricao = models.CharField(max_length=255)
     preco = models.DecimalField(max_digits=10, decimal_places=2)
     acabamento = models.ForeignKey("Acabamento", on_delete=models.PROTECT)
-    abatimento_mm = models.IntegerField(default=0, help_text="Abate em mm na metragem do cálculo")
+    abatimento_mm = models.DecimalField(max_digits=6, decimal_places=2, default=0, help_text="Abate em mm na metragem do cálculo")
     modelo = models.CharField(max_length=50, blank=True, null=True)
     bimer_id = models.CharField(
         max_length=20, blank=True,
